% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.Normalized.Number.Of.Hits}
\alias{get.Normalized.Number.Of.Hits}
\title{To obtain the number of hits during 'years', in database='db'}
\usage{
get.Normalized.Number.Of.Hits(years = 1980:2014, query = "",
  db = "pubmed", normalization_value = 1e+06)
}
\arguments{
\item{x}{}
}
\value{
y
}
\description{
To obtain the number of hits during 'years', in database='db'
}
\author{
HK
}

