% Generated by roxygen2 (4.0.2): do not edit by hand
\name{informationRetrieval}
\alias{informationRetrieval}
\title{This is basically a test funtion and will print only a message}
\usage{
informationRetrieval()
}
\arguments{
\item{NULL}{}
}
\value{
A string: a message
}
\description{
This is basically a test funtion and will print only a message
}
\author{
Hashem Koohy
}

