% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Power}
\alias{Power}
\title{This is basically a test funtion and will print only a message}
\usage{
Power(x, n)
}
\arguments{
\item{x,}{n}
}
\value{
x to power of n
}
\description{
This is basically a test funtion and will print only a message
}
\author{
Hashem Koohy
}

